MOVE mqtt struct
