--- e2l-parser/src/mqtt_client.rs
@@ -1,1 +1,0 @@
-pub mod mqtt_structs;
